move topic controllers into infrastructure folder
