+++ src/Infrastructure/KafkaJanitor.RestApi/Features/Topics/Infrastructure/AccessController.cs
@@ -4,7 +4,7 @@ using KafkaJanitor.RestApi.Features.ServiceAccounts.Infrastructure;
 using KafkaJanitor.RestApi.Features.Topics.Domain.Models;
 using Microsoft.AspNetCore.Mvc;
 
-namespace KafkaJanitor.RestApi.Features.Topics
+namespace KafkaJanitor.RestApi.Features.Topics.Infrastructure
 {
     [Route(Routes.ACCESS_ROUTE)]
     public class AccessController : ControllerBase
