+++ src/Infrastructure/KafkaJanitor.RestApi/Features/Topics/Infrastructure/TopicsController.cs
@@ -1,9 +1,8 @@
 using System.Threading.Tasks;
-using KafkaJanitor.RestApi.Features.Topics;
 using KafkaJanitor.RestApi.Features.Topics.Domain.Models;
 using Microsoft.AspNetCore.Mvc;
 
-namespace KafkaJanitor.RestApi.Features.Topics
+namespace KafkaJanitor.RestApi.Features.Topics.Infrastructure
 {
     [Route(Routes.TOPICS_ROUTE)]
     public class TopicsController : ControllerBase
