+++ src/Infrastructure/KafkaJanitor.RestApi/Startup.cs
@@ -2,6 +2,7 @@ using KafkaJanitor.RestApi.Enablers.Metrics;
 using KafkaJanitor.RestApi.Features.AccessControlLists.Infrastructure;
 using KafkaJanitor.RestApi.Features.ServiceAccounts.Infrastructure;
 using KafkaJanitor.RestApi.Features.Topics;
+using KafkaJanitor.RestApi.Features.Topics.Infrastructure;
 using Microsoft.AspNetCore.Builder;
 using Microsoft.AspNetCore.Hosting;
 using Microsoft.Extensions.Configuration;
